moving currentTrade from gameState to playerState
