+++ client/store/actions/gameState.js
@@ -1,10 +1,6 @@
 import {
   nextPlayer,
   SET_GAME_USERS,
-  MAKE_OFFER,
-  RECEIVE_OFFER,
-  ACCEPT_OFFER,
-  REJECT_OFFER,
   distributeResource,
   distributeResourcePlayer,
   rollDice,
@@ -77,36 +73,6 @@ export const newDiceRoll = () => {
       socket.emit('dispatchThunk', {action: 'robberThunk'})
     }
   }
-}
-
-export const makeOffer = currentTrade => {
-  socket.emit('dispatch', {
-    type: RECEIVE_OFFER,
-    currentTrade
-  })
-  return {type: MAKE_OFFER, currentTrade}
-}
-
-export const receiveOffer = currentTrade => {
-  return {type: RECEIVE_OFFER, currentTrade}
-}
-
-export const acceptOffer = playerNumber => {
-  console.log('acceptOffer', playerNumber)
-  socket.emit('dispatch', {
-    type: ACCEPT_OFFER,
-    playerNumber
-  })
-  return {type: ACCEPT_OFFER, playerNumber}
-}
-
-export const rejectOffer = playerNumber => {
-  console.log('rejectOffer', playerNumber)
-  socket.emit('dispatch', {
-    type: REJECT_OFFER,
-    playerNumber
-  })
-  return {type: REJECT_OFFER, playerNumber}
 }
 
 export const moveRobberThunk = id => (dispatch, getState) => {
